# openapi.yml
openapi: 3.0.0
info:
  title: Track My Tasks API
  description: |
    API specification for Track My Tasks service.
  version: 0.0.0
  contact:
    name: Ваше имя или команда
    email: support@example.com
    url: https://example.com
  license:
    name: AGPL-3.0
    url: https://opensource.org/licenses/AGPL-3-0
servers:
  - url: https://api.trackmytasks.org/v1
    description: Public SaaS instance
  - url: https://staging-api.trackmytasks.org/v1
    description: Staging instance
  - url: http://localhost:8080/v1
    description: Development instance
...
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks
      description: Returns paginated collection of tasks
      operationId: getTasks
      parameters:
        - $ref: "#/components/parameters/PaginationPage"
        - $ref: "#/components/parameters/PaginationLimit"
        - $ref: "#/components/parameters/SortingOrder"
        - in: query
          name: sort-field
          schema:
            $ref: "#/components/schemas/TaskSortableField"
          description: By which field results are sorted
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
              examples:
                default:
                  summary: Empty response example
                  value:
                    data: []
                    meta:
                      total: 0
                      page: 1
                      limit: 20
                      totalPages: 0
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: "#/components/responses/Forbidden"
        '405':
          $ref: "#/components/responses/MethodNotAllowed"
        '429':
          $ref: "#/components/responses/RateLimited"
        '500':
          $ref: "#/components/responses/ServerError"

    post:
      tags:
        - Tasks
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskBody'
            examples:
              default:
                summary: Example of creation of typical task
                value:
                  a: 123
      responses:
        '201':
          description: Resource created
          headers:
            Location:
              schema:
                type: string
              description: URL of the created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: "#/components/responses/Forbidden"
        '405':
          $ref: "#/components/responses/MethodNotAllowed"
        '429':
          $ref: "#/components/responses/RateLimited"
        '500':
          $ref: "#/components/responses/ServerError"
  /tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get task by id
      operationId: getTaskById
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: Id of the task
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: "#/components/responses/Forbidden"
        '405':
          $ref: "#/components/responses/MethodNotAllowed"
        '429':
          $ref: "#/components/responses/RateLimited"
        '500':
          $ref: "#/components/responses/ServerError"

    put:
      tags:
        - Tasks
      summary: Update the task
      operationId: updateTask
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskBody'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: "#/components/responses/Forbidden"
        '405':
          $ref: "#/components/responses/MethodNotAllowed"
        '429':
          $ref: "#/components/responses/RateLimited"
        '500':
          $ref: "#/components/responses/ServerError"

    delete:
      tags:
        - Tasks
      summary: Remove the task
      operationId: deleteTask
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Resource removed
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: "#/components/responses/Forbidden"
        '405':
          $ref: "#/components/responses/MethodNotAllowed"
        '429':
          $ref: "#/components/responses/RateLimited"
        '500':
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    PaginationPage:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PaginationLimit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of elements in returned collection
    SortingOrder:
      in: query
      name: sort
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: By which order results are sorted
  schemas:
...
    Task:
      allOf:
        - $ref: "#/components/schemas/IdentifiableResource"
        - $ref: "#/components/schemas/TaskBody"
    TaskBody:
      type: object
      required:
        - status
      properties:
        template_id:
          type: integer
          description: Id of the template from which task was created
          example: 3827
        moved_task_id:
          type: integer
          description: Id of the task from which this task was moved
          example: 8572
        status:
          type: string
          enum: ['pending', 'running', 'finished', 'cancelled', 'moved', 'overdue']
          description: Current state of the task
        notes:
          type: string
          minLength: 2
          description: Additional notes about the task
        date:
          type: string
          format: date
          description: Date that the task was assigned at
        started_at:
          type: string
          format: date-time
          description: Timestamp when task was started
        ended_at:
          type: string
          format: date-time
          description: Timestamp when task was ended
    TaskSortableField:
      type: string
      enum: [id, template_id, moved_task_id, status, notes, date, started_at, ended_at]
      default: id
...
    PaginationMeta:
      type: object
      required:
        - total
        - page
        - limit
        - total_pages
      properties:
        total:
          type: integer
          description: Total number of elements
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Maximum number of elements per page
        total_pages:
          type: integer
          description: Total number of available pages
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    IdentifiableResource:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          description: Unique resource identifier
          # examples:
          #   - 1
          #   - 18429485572
  responses:
    BadRequest:
      description: Bad request returned if client provided data in wrong format.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    Unauthorized:
      description: Not authorized returned if client need to authenticate to access the endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden returned if client does not have permissions to access the endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    MethodNotAllowed:
      description: Method not allowed returned if client tries to access the endpoint using wrong http method.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
            # todo: allow header
    RateLimited:
      description: Too many requests returned if client ran out of request quota.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Unexpected internal error returned if something went wrong at server side.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'